Remove go-pkg template (#80)
* feat: remove go-pkg template that is now deprecated

* chore: remove go-pkg template from flake-language.yml and update README

- Removed the `go-pkg` template from the `.github/workflows/flake-language.yml` file.
- Updated the README to remove the reference to `go-pkg` for building legacy Go apps.

* chore: remove go-pkg from flake-language.yml configuration

* chore: update home-manager and nixpkgs to release-24.11

Updated the URLs for home-manager and nixpkgs in the flake.nix files for darwin, nixos-desktop, and nixos-minimal templates to point to the release-24.11 versions. This ensures the templates use the latest versions of these dependencies.

* chore(configuration): remove redundant sound.enable setting

The sound.enable setting has been removed as it is redundant with the current configuration. Pulseaudio is disabled and Pipewire is being used for sound management.

* chore: update home-manager and nixpkgs to release-25.05

Updated the home-manager and nixpkgs URLs in the flake.nix file to point to the release-25.05 versions. This ensures compatibility with the latest features and improvements in the new release.

* chore: update flake.nix to use latest home-manager and nixpkgs

Updated the home-manager URL to point to the latest version instead of a specific release. Changed the nixpkgs URL to use the nixos-unstable branch for more up-to-date packages.

* chore: update stateVersion to 24.11

- Updated `system.stateVersion` and `home.stateVersion` across various modules to "24.11".
- Ensures consistency in versioning for system and home-manager configurations.
- Affects darwin, nixos-desktop, and nixos-minimal templates.

* feat(nix): enable automatic optimization

- Added `optimise.automatic = true;` to the Nix configuration in darwin, nixos-desktop, and nixos-minimal templates.
- Removed `auto-optimise-store = true;` from the Nix settings as it is now replaced by the new automatic optimization setting.

* refactor(darwin): remove nix-daemon service enablement

The nix-daemon service enablement has been removed from the darwin.nix configuration. This change simplifies the module configuration by eliminating the explicit setting of services.nix-daemon.enable to true.